Gardening: Remove unnecessary synchronization in LoggingMeter
Motivation
----------
LoggingMeter.valueRecorder() does not need to be synchronized,
because valueRecorders is a ConcurrentMap, and
ConcurrentMap.computeIfAbsent is an atomic operation.

LoggingMeter.Worker.dumpMetrics() does not need to be synchronized,
because it can only be called by the metrics worker thread.

Note: At first glance, it might look like the synchronization
prevented dumpMetrics() from running at the same time
as valueRecorder(), but that's not the case! These are methods
of two different classes, so the sync was independent.
(Even if that was the original intent, it's unnecessary.)

Modifications
-------------
Remove synchronized keyword from LoggingMeter.valueRecorder()
and LoggingMeter.Worker.dumpMetrics().

Change the type of `valueRecorders` from Map to ConcurrentMap.
This does not affect the code's behavior, but it gives readers
another clue that valueRecorders.computeIfAbsent() is atomic.

Change-Id: I1e3e67247ca7846bde5b753099abac78b2d16491
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219596
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>