JVMCBC-1602 App Telemetry
Modifications
-------------
Added volatile environment config properties:

* appTelemetryEndpoint
    default = null (use endpoint from topology)

* disableAppTelemetry
    default = false (capture telemetry)

Added AppTelemetryCollector for gathering telemetry,
and AppTelemetryReporter for publishing it via WebSocket.

Added AppTelemetry, which holds a collector and reporter.
An instance is initialized in Core's constructor.

Core has a new `appTelemetryCollector()` accessor.
(Decided against adding this to CoreResources,
because it's not relevant for Protostellar.)

Request latency is captured in RequestContext.dispatchLatency()
when the request terminates.

Cancellation due to closed channel is captured in
RetryOrchestrator.maybeRetry().

All other cancellations and timeouts are captured in
BaseRequest.cancel().

Update MockUtil to mock ApplicationTelemetryCollector.

Change-Id: I06769913306f987fd237acce056aaf91a5499d65
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/219666
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>