JVMCBC-1592 Shuffle node list instead of rotating
Motivation
----------
Avoid a pathological condition where the number of open buckets
is equal to the number of nodes, and the config refresh request
for one of the buckets always goes to a degraded node.

By randomizing the node selection, we ensure the request
eventually goes to one of the other nodes.

Change-Id: I013328fab5286775d796b1cc57ca0ba7933bb460
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/221189
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>