JVMCBC-1589 Adopt JSpecify in core.topology
Modifications
-------------
Added dependency on JSpecify. The JSpecify experts
explicitly recommend using the default scope ("compile"),
and *not* hiding it by using "optional" or "provided" scope.

Replaced Reactor nullability annotations with JSpecify annotations.

Added missing @Nullable annotation to AutoNetworkSelector.cachedResult.

Moved some @Nullable annotations closer to the annotated types.
This isn't strictly necessary, but is a good convention.

Annotated CbObjects.defaultIfNull() as @NullUnmarked, because
the nullability of the return value depends on the nullability
of the default value. Might revisit this later.

Change-Id: I6aa8cc324a70f42c35d7d0dad227549a70bbf4a5
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/221304
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>