JVMCBC-1597 Populate ClusterConfig using new ClusterTopology model
Motivation
----------
Migrate more code to the new ClusterTopology model.

Modifications
-------------
Change `ClusterConfig.setGlobalConfig()` to accept
`ClusterTopology` instead of `GlobalConfig`.

Change `ClusterConfig.setBucketConfig()` to accept
`ClusterTopologyWithBucket` instead of `BucketConfig`.

This is a minor change to production code, with a
disproportionately large impact on unit tests
that create mock topologies.

Add a ClusterTopologyBuilder that makes it easy for
unit tests to craft special topologies for testing,
without having to resort to mocks.

Change-Id: Ic53b93408f0b9a40e2b2b160b34e62c7a26cd99d
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/221639
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>