JVMCBC-1596 (followup) Fix broken FIT performer build
Motivation
----------
The change in 40dfd93892 to use `maven.compiler.release`
instead of `maven.compiler.source/target` broke the
FIT performer validation process, because it was defaulting
to a too-old version of the maven-compiler-plugin.

We should look into whether that process is correctly
using this project's Maven wrapper, but in the meantime...

Modifications
-------------
Specify the Maven compiler plugin version.

Change-Id: I47b47b6577381d3331c84c8e32af879b0ead28ce
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/222162
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>