JVMCBC-1603 Parse `nodeUUID` from cluster topology
Motivation
----------
Groundwork for application telemetry.

Modifications
-------------
Parse `nodeUUID` from app telemetry path
and add it to `NodeIdentifier`.

If `nodeUUID` is not present in topology JSON,
create a synthetic one from the node's canonical
host and port, which was the best way to uniquely
identify a node prior to Couchbase Server 8.0.0.

Change-Id: I7bf271218575d1c7b58708c1c0248e335677b6d8
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/222235
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>