KCBC-180 KotlinxSerializationJsonSerializer cannot serialize or deserialize null
Motivation
----------
The experimental KotlinxSerializationJsonSerializer class
fails to serialize or deserialize JSON documents that have
null at the root. This includes full Couchbase documents,
as well as SQL++ query parameters and result rows.

Analysis
--------
Getting full Kotlin type information from typeOf<T>() is expensive
in Kotlin 1.6, so we've been making do with the plain old Java type.
The Java type doesn't have nullability info. When we use it to get
a serializer, we receive a serializer for a non-null type,
regardless of whether the Kotlin type was nullable.
This non-null serializer rejects nulls, even if the original Kotlin
type is nullable.

Modifications
-------------
When acquiring a type's KSerializer, always adjust it to be nullable.

Whe deserializing, add a root-level null check if we know the type
is non-nullable.

A null check is not required when serializing, because Kotlin's type system
prevents passing inappropriate nulls in that context.

Enabled related tests that had previously been disabled.

Change-Id: If39f7cbe0f4abf5403cba09f629a260ee861b682
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/222632
Reviewed-by: G. Blake Meike <blake.meike@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>