JVMCBC-1615 Move nodeUUID from NodeIdentifier to HostAndServicePorts
Motivation
----------
A NodeIdentifier created during bootstrap does not
have a UUID, even if the server supports node UUIDs.

These UUID-less NodeIdentifiers can stick around
because they're baked into the initial `Node` objects,
and stay there after we fetch the cluster topology
that has the real UUID info.

Having UUID be a property of NodeIdentifier
made it too easy to assume the property was accurate
when it fact it wasn't. This was the source of a bug
in the initial application telemetry implementation.

Modifications
-------------
Store the `nodeUUID` value in `HostAndServicePorts`
instead of `NodeIdentifier`. This way it's clear we
don't know the UUID until we have the cluster topology.

Make the UUID property nullable instead of "helpfully"
defaulting to the node's canonical address if UUID is absent;
we don't know if there will be a future use case that
requires knowing if the UUID is "real" or synthetic.

Change-Id: I41d734faa38827062832dc025139717f81bc40ce
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/223247
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>