KCBC-189 Promote KotlinxSerializationJsonSerializer
Motivation
----------
kotlinx.serialization is popular in the Kotlin ecosystem.

Modifications
-------------
Add a `ktype` field to TypeRef, and use it to get a KSerializer
the "right way" in KotlinxSerializationJsonSerializer.

This comes with a small performance penalty for developers
using other JsonSerializer implementations. To mitigate this,
optimize typeRef() for pre-serialized content so power users
can squeeze every nanosecond out of serialization times.

Change-Id: Icbd256aa108b5b46f0070d847f744df88e3325ab
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/224958
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: G. Blake Meike <blake.meike@couchbase.com>