JVMCBC-1645 Test: Stabilize QueryDelayDisconnectIntegrationTest
Motivation
----------
This test was failing on CI because it was not letting
the base class configure the cluster environment
via `createCluster()`.

Modifications
-------------
Create the cluster using `createCluster()` so the
base class can configure the cluster environment.

Rework the test so it more reliably triggers the
`EndpointDisconnectDelayedEvent`.

Unfortunately, there doesn't seem to be a great
way to cancel an in-flight query, so the test relies
on a long-running query that is expected to time out.

Change-Id: Ief262d95be64fc40e1d35b837a1a64e0dd6e102f
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/228014
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>