JVMCBC-1650 Backport Maven Central Portal publishing changes
Motivation
----------
Align the `aurora` branch with changes in master
to support publication via Maven Central Portal
and GitHub Actions.

Unfortunately, could not cherry-pick due to merge conflicts.

Modifications
-------------
Backport the following patchsets:

https://review.couchbase.org/c/couchbase-jvm-clients/+/226920
https://review.couchbase.org/c/couchbase-jvm-clients/+/227011
https://review.couchbase.org/c/couchbase-jvm-clients/+/228347
https://review.couchbase.org/c/couchbase-jvm-clients/+/228435
https://review.couchbase.org/c/couchbase-jvm-clients/+/228480

The same changes are applied verbatim, with two exception:

1. In `aurora`, the parent POM's `scala-2.13` profile sets
`scala.compat.library.version` to 2.13.15 to match the
scala-client 1.8.1 release. This differs from the `master`
branch, which upgrades that library to 2.13.16.

2. The "Deploy Snapshot" action in the `aurora` branch
triggers on pushes to `aurora`, not `master`.

Change-Id: I61e8d3519eec06c3e3de467ac4bffcb7b4dda71d
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/228481
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>