TestFix: KeyValueCollectionIntegrationTest.recognizesCollectionAfterCreation
Motivation
----------
This test was failing against Community Edition
because it used the deprecated createCollection(CollectionSpec)
method which sends the enterprise-only `maxTTL` parameter
regardless of whether the caller specified it.

Modifications
-------------
Update the deprecation warning to indicate the method
is not compatible with Community Edition.

Update the test to use the non-deprecated version
of `createCollection`, which works with CE.

Rejected Alternatives
---------------------
We _could_ change the deprecated method
to convert a zero TTL into null so it wouldn't
get sent to the server. However, this would ignore
a zero TTL explicitly set by the user, breaking
the behavior if the server's default TTL ever changes
to something other than zero.

Change-Id: I3fac794436b71254b68bcf39ec444702493a2a87
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/228634
Reviewed-by: Will Broadbelt <william.broadbelt@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>