JVMCBC-1657 Better backpressure for row-based HTTP services
Motivation
----------
`autoRead` was activated every time an item was requested,
even if there was an existing surplus of buffered rows.
This could lead to reading and buffering more rows than the
consumer requested, potentially without limit.

The problem was most pronounced when result rows are small
enough that several fit in a single network receive buffer.

Modifications
-------------
Rename `requested` to `demand`, because the value
decreases as demand is met (when rows are supplied).

Instead of enabling autoRead whenever an item is requested,
enable it only if demand is positive.

Prevent a race condition by synchronizing access
to the `demand` counter and the autoRead setting.

Change-Id: Ia82ed52a7182b380efaec81d3a70812d21d340df
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/229015
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>