JVMCBC-1656 Java SDK Seed Nodes Don't Always Get Correctly Updated
Motivation
----------
In scenarios where the global topology loader fails,
we want to get global topology info from the buckets.

Modifications
-------------
Take advantage of the relationship between bucket and
global topology: a bucket topology IS A global topology
with extra info specific to the bucket.

Merge the code paths for apply proposed topology updates,
so global and bucket topologies go in the same funnel.

Use the global part of a bucket topology to update
the SDK's view of the global topology.

Simplify `updateSeedNodeList`, because now it only
needs to look at the current global topology.

Change-Id: Ie6a6b8e157bf85891815a79eb2de9885ffe80c15
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/229025
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>