Java FIT performer: support two ClusterEnvironment modes
3.2.6 added an easy way for SDK users to configure
the SDK without having to take ownership of
ClusterEnvironment management.  It also allows
passing parameters in the connection string, which
is not allowed with those externally owned
ClusterEnvironments.

The performer will now use the "new" mode where
possible.  The features required by the new mode
are likely not to be required in the situations
where we are using the old mode.

Change-Id: I58755c369856bd977779fee4174ca54814189dea
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/229357
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>