JVMCBC-1660 Change Network Heuristic for JVM SDKs
Motivation
----------
Consistent behavior between SDKs.

Make certain private endpoint deployments work
without the user having to explicitly specify
`network=external`.

Modifications
-------------
When there's no exact match between a seed node and
an address in the cluster topology, fall back to
"external" network (if present) instead of "default".

Throw an exception if the user explicitly specifies
a non-existent network. This results in a `ConfigIgnoredEvent`
in the log at WARN level with a message that includes:

"""
Requested network 'external' is not available for this cluster.
Available networks: [default]
"""

CAVEATS
-------
This is potentially a breaking behavior change.
Upgrade with caution.

Change-Id: I38675cd02456088662c29a4fa315720a651e65c1
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/229917
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Build Bot <build@couchbase.com>