SCBC-378: Support Scala 3
The Scala 3 version of the SDK, while built from the same code-base,
diverges significantly to the point that it should be considered a
new major.

These major changes are made:

* A new Scala 3.3 build of the SDK is now to be published, alongside
  existing 2.12 and 2.13 builds.  The 3.3 build is compatible with apps
  compiled with 3.3 to 3.7 inclusive, plus 2.13.  It is now the primary
  Scala SDK build.
* Reactor API is not carried forward, as the Scala Reactor Extensions library
  that it depends on is long-EOL.
* Not carrying forward built-in support for JSON libraries (Circe,
  json4s etc).
* Not carrying forward support for views, analytics, replica reads, all management
  operations, and datastructures.
* There are two overloads for most API methods, one that takes an options
  block, and one that tries to be a convenience method offering a few popular
  settings.  The latter is removed, significantly tidying the API.

All this functionality continues to work unchanged in the Scala 2 builds.
(With the small exception of the removal of a long-deprecated transactions
method).

Support for 2.11 has been removed.

scalafmt is temporarily removed.  Since this change touches almost
every file in the SDK, a follow-up commit is going to scalafmt
the full SDK.

Testing
=======
All tests pass for both Scala 2 and 3 builds.
https://github.com/couchbaselabs/scala-sdk-test-matrix created to test the matrix
of supported combinations.
The FIT performer continues to exclusively test the 2.12 build.

Change-Id: I5de11eae460fefc91144d830c0a0699af3980c03
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/230429
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>