Add a one-off GHA workflow for publishing from master
Motivation
----------
In tag 3.9.0, the scala-3 client is broken.
This prevented the scala-3 client from being published,
and also prevented the API reference (Javadoc / Scaladoc)
for all modules from being published to S3.

Modifications
-------------
Add a one-off GHA workflow that publishes a release from
the `master` branch. To prevent accidentally running this
workflow, we'll delete it after publishing 3.9.0.

Configure `central-publishing-maven-plugin to
ignore existing published components, since some of the
artifacts were already published. This is good for the
future as well, in case we need to re-run a release
workflow due to a transient failure.

Change-Id: I107ba8c649533f56338d84eb7f3273ec83661577
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/232116
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>