JVMCBC-1686 CertificateAuthenticator.toString() can include unprotected private key
Motivation
----------
Prior to this change, the toString() method of an authenticator
created by CertificateAuthenticator.fromKey() includes the private key.
This could potentially result in the private key appearing in
application logs -- but only if the user explicitly logs the authenticator.

Modifications
-------------
Change CertificateAuthenticator.toString() to return just
"CertificateAuthenticator".

Omit `keyManagerFactory` because it's a supplier whose
toString() is not interesting.

Omit `keyCertChain` because it's not particularly useful.

Change-Id: I733830eb8784f9c0195074903544ee8eb42ce334
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/233953
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>