JVMCBC-1689 Omit redundant authorization ID from OAUTHBEARER payload
Motivation
----------
Now that MB-68594 is resolved, the SDK no longer needs to
redundantly specify the authorization ID when using a JWT
to authenticate.

Modifications
-------------
Remove code that extracted the "sub" field from the JWT,
since we no longer need to pass it to the server redundantly.

Let subclasses of InitialResponseSaslClient decide
which bits of information to fetch from the callback.
This lets us specify a null username, while avoiding
the exception that would result from trying to fetch
a null username from the callback.

Propagate the authorization ID from the SaslClientFactory
to OauthBearerSaslClient. This feature is currently unused,
(the whole point of this change is to avoid having
to explicitly pass an authorization ID) but this parallels
the other SaslClient implementations.

When authorization ID is absent (currently always!),
omit the "a=<authorization-id>" component from the SASL
OAUTHBEARER payload.

Change-Id: I606f8297cc5bc6547f7ef5b1adb276eaa41333b0
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/234802
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>