JVMCBC-1693 Debounce GetConfig requests
Motivation
----------
During rebalance, the SDK receives many cluster topology
change notifications. Instead of fetching the new topology
once for each notification, the SDK should reduce the number
of fetches by ignoring obsolete notifications.

Modifications
-------------
Process topology update notifications one at a time:

* Serialize the topology update signal processing by using
`concatMap` instead of `flatMap`.

* Before fetching new topology, compare current topology against
the latest version the server told us about. If we have up-to-date
topology, ignore the obsolete notification.

* Call `provider.propose{Bucket,Global}Config` in `doOnNext`
instead of in an external subscriber. This ensures the SDK's
"current" topology revision is updated immediately, before
processing the next signal.

--

Add `TopologyChangeNotificationBuffer` for storing notifications.
Only ask server for latest topology if a notification is still relevant.

Use this separate buffer for the topology revisions because
storing them in a Flux would be excessively complex.

--

Move the topology polling trigger flux creation code into
DefaultConfigurationProvider, so it can be shared by the
Global and KV refreshers. (The flux itself isn't shared --
only the code to create it).

For consistency, and to ensure we don't block the KV threads,
handle topology change notifications in the same thread as
timer ticks (instead of in the KV thread).

Change-Id: Icabc7b10eabea6977a376b931f4e4c43f05350bb
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/235174
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>