JVMCBC-1696 Should not make bucketful KV connections to nodes that aren't hosting the bucket
Motivation
----------
During the migration to the new cluster topology classes,
there was a regression that caused `core.reconfigureBuckets`
to create bucket-scoped KV endpoints for nodes that
aren't hosting the bucket.

During rebalance, this causes noise in the logs due to fruitless
connection attempts, and more importantly causes `Cluster.diagnostics()`
to report the cluster is in a degraded state because the doomed
endpoints are not connected.

Modifications
-------------
When reconfiguring a bucket, transform the nodes by removing
the KV service from any node that is not hosting the bucket.
This echoes the behavior of the old BucketConfig class.

Note that we need to process the whole node list from the
"global" part of the topology so we can remove existing
KV endpoints that are no longer required by the updated topology.

Change-Id: I8eaf79d39033d4a24cfdd5b1eb18864d296b931e
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/235828
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>