JVMCBC-1700 Upgrade to Jackson 2.20.1
Modifications
-------------
Use Jackson BOM.

In scala-client POM, remove explicit dependencies
on jackson-core and jackson-annotations; these are
transitive dependencies of jackson-databind, and
are included automatically.

Also noteworthy, starting with Jackson 2.20.0,
the jackson-annotations library version is just `2.20`
(no patch version). This isn't a problem if we
rely on the BOM (or rely on transitive dependency
inclusion instead of explicitly depending on a specific
version).

Change-Id: Ife4289558c507b93b85d3c8b1dfa0e8597e8686f
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/236250
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>